fix vnpay

+++ lib/config/shared_preferences.dart
@@ -21,11 +21,11 @@ class SharedService {
   static void setTheme(bool isDark) {
     _pref.setBool("is_dark", isDark);
   }
-  static void setFont(int fontSize){
-    _pref.setInt('font', fontSize);
+  static void setFont(double fontSize){
+    _pref.setDouble('font', fontSize);
   }
-  static int? getFont(){
-    return _pref.getInt("font");
+  static double? getFont(){
+    return _pref.getDouble("font");
   }
   static void clear(){
     _pref.clear();
